Moved the category styling and sorting to the view model
